Move backend fetching to a separate module
